Move lib into src directory to match Wanda project structure
