+++ src/lib/js/core.js
@@ -1,12 +1,12 @@
 import equal from "fast-deep-equal/es6/index.js";
-import { Module, makeModule } from "../runtime/Module.js";
-import { Cons, cons } from "../shared/cons.js";
-import { print } from "../printer/print.js";
-import { println } from "../printer/println.js";
-import { printString } from "../printer/printString.js";
-import { isTruthy, makeKeyword } from "../runtime/utils.js";
-import { Exception } from "../shared/exceptions.js";
-import { hasMetaField } from "../runtime/object.js";
+import { Module, makeModule } from "../../js/runtime/Module.js";
+import { Cons, cons } from "../../js/shared/cons.js";
+import { print } from "../../js/printer/print.js";
+import { println } from "../../js/printer/println.js";
+import { printString } from "../../js/printer/printString.js";
+import { isTruthy, makeKeyword } from "../../js/runtime/utils.js";
+import { Exception } from "../../js/shared/exceptions.js";
+import { hasMetaField } from "../../js/runtime/object.js";
 
 // Module name
 const name = "Core";
